% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wor_corpus}
\alias{wor_corpus}
\title{War of Rebellion Corpus.}
\format{A list with 129 entries, each containing a char vector of the documents in their raw text form.}
\source{
\url{https://ehistory.osu.edu/books/official-records/index}
}
\usage{
wor_corpus
}
\description{
The text contents of each page in the corpus. A list of the serials, where each entry contains a character vector of all the documents in that serial. Note that as in the info dataframe, serials 112 and 113 are missing, in this object they are empty.
}
\keyword{datasets}
