% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wor_info}
\alias{wor_info}
\title{War of Rebellion Info.}
\format{A data frame with five variables: \code{SERIES}, \code{VOLUME}, \code{PART}, \code{SERIAL}, \code{CAMPAIGN}.}
\source{
\url{https://ehistory.osu.edu/books/official-records/index}
}
\usage{
wor_info
}
\description{
Information about the source and contents of each serial in the War of Rebellion corpus hosted online at Ohio State University. Values in the \code{SERIAL} column correspond with the entries in the wor_corpus list. Note that serials 112 and 113 are missing from the web source.
}
\keyword{datasets}
